Add init version of e2e test structure
