+++ e2e_tests/README.md
@@ -1,0 +1,15 @@
+# End-to-End Tests
+
+## Structure
+
+- `apis/`: Contains the E2E test cases for different APIs.
+- `utils/`: Contains utility functions for the tests.
+- `conftest.py`: Pytest configuration file.
+- `testdata/`: Contains test data and configuration files.
+
+## Adding new end-to-end tests
+
+1. Create a new test file in the `apis/` directory.
+2. Use the existing utility functions from the `utils/` directory.
+3. Follow the naming convention and structure of existing test files.
+4. Add any necessary configuration files to the `testdata/` directory.
