+++ e2e_tests/apis/testdata/valid_dockerhub_config.yaml
